pyproject.toml
[project]
name = "clip-playground"
version = "0.1.0"
description = "Minimal CLIP web demo with Flask."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0,<4.0",
    "httpx>=0.28.1",
    "pillow>=10.0",
    "socksio>=1.0.0",
    "torch>=2.2",
    "transformers>=4.38",
]

